Move TSimpleXxx sensors from library/profiling/solomon to library/profiling commit_hash:e24d2de3bb12bd63777f8ac6acdb48bcbad259b1
