+++ yt/yt/library/profiling/solomon/helpers.cpp
@@ -4,6 +4,7 @@
 #include "sensor_set.h"
 
 #include <yt/yt/library/profiling/percpu.h>
+#include <yt/yt/library/profiling/sensor_impl.h>
 
 #include <yt/yt/core/http/http.h>
